requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
fake-useragent==2.2.0
google-auth==2.43.0
google-genai==1.52.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.3.5
pillow==12.0.0
playwright==1.56.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycurl==7.45.7
pydantic==2.12.4
pydantic_core==2.41.5
pyee==13.0.0
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
python-dotenv==1.2.1
random-user-agent==1.0.1
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
undetected-playwright==0.3.0
urllib3==2.5.0
websockets==15.0.1
